codizer/python_backend/requirements.txt
django>=5.2.0
djangorestframework>=3.16.0
ast2json>=0.4
big-o>=0.11.0
numpy>=2.2.5 